backend/requirements.txt
## requirements
fastapi
uvicorn
pymongo
python-dotenv
pytest